MAIN: Round datetimes to nearest multiple of 10
round_datetime() takes a datetime and sets the minutes to the nearest
multiple of ten (i.e. 00, 10, 20, 30, 40, 50).

The smallest interval the Steam Web API offers for its playtime data
is about 10 minutes. Saving playtime data in smaller intervals than 10
minutes is therefore pointless. One solution to this problem is to
adjust the logging time, which round_datetime() is used for.